splitup raster_layer
