+++ src/actinia_core/rest/raster_layer.py
@@ -4,7 +4,7 @@
 # performance processing of geographical data that uses GRASS GIS for
 # computational tasks. For details, see https://actinia.mundialis.de/
 #
-# Copyright (c) 2016-2021 Sören Gebbert and mundialis GmbH & Co. KG
+# Copyright (c) 2016-2022 Sören Gebbert and mundialis GmbH & Co. KG
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -31,24 +31,21 @@ import pickle
 from uuid import uuid4
 from werkzeug.utils import secure_filename
 from actinia_api.swagger2.actinia_core.schemas.raster_layer import \
-     RasterInfoResponseModel, RasterInfoModel
+     RasterInfoResponseModel
 
 from actinia_core.core.common.app import URL_PREFIX
-from actinia_core.processing.actinia_processing.ephemeral.ephemeral_processing \
-     import EphemeralProcessing
-from actinia_core.processing.actinia_processing.ephemeral.persistent_processing \
-     import PersistentProcessing
 from actinia_core.rest.base.map_layer_base import MapLayerRegionResourceBase
 from actinia_core.core.common.redis_interface import enqueue_job
-from actinia_core.core.common.exceptions import AsyncProcessError
 from actinia_core.core.utils import allowed_file
 from actinia_core.models.response_models import \
     ProcessingResponseModel, ProcessingErrorResponseModel
 from actinia_core.models.response_models import SimpleResponseModel
+from actinia_core.processing.common.raster_layer import \
+     start_info_job, start_delete_job, start_create_job
 
 __license__ = "GPLv3"
 __author__ = "Sören Gebbert, Carmen Tawalika, Guido Riembauer, Anika Weinmann"
-__copyright__ = "Copyright 2016-2021, Sören Gebbert and mundialis GmbH & Co. KG"
+__copyright__ = "Copyright 2016-2022, Sören Gebbert and mundialis GmbH & Co. KG"
 __maintainer__ = "mundialis"
 
 
@@ -276,172 +273,3 @@ class RasterLayerResource(MapLayerRegionResourceBase):
 
         http_code, response_model = pickle.loads(self.response_data)
         return make_response(jsonify(response_model), http_code)
-
-
-def start_info_job(*args):
-    processing = EphemeralRasterInfo(*args)
-    processing.run()
-
-
-class EphemeralRasterInfo(EphemeralProcessing):
-
-    def __init__(self, *args):
-
-        EphemeralProcessing.__init__(self, *args)
-
-    def _execute(self):
-        """Read info from a raster layer
-
-        Use a temporary mapset for processing
-        """
-        self._setup()
-
-        raster_name = self.map_name
-        self.required_mapsets.append(self.mapset_name)
-
-        pc = {}
-        pc["1"] = {
-            "module": "r.info", "inputs": {
-                "map": raster_name + "@" + self.mapset_name},
-            "flags": "gre"}
-
-        self.skip_region_check = True
-        process_list = self._create_temporary_grass_environment_and_process_list(
-            process_chain=pc, skip_permission_check=True)
-        self._execute_process_list(process_list)
-
-        kv_list = self.module_output_log[0]["stdout"].split("\n")
-
-        raster_info = {}
-
-        for string in kv_list:
-            if "=" in string:
-                k, v = string.split("=", 1)
-                raster_info[k] = v
-
-        self.module_results = RasterInfoModel(**raster_info)
-
-
-def start_delete_job(*args):
-    processing = PersistentRasterDeleter(*args)
-    processing.run()
-
-
-class PersistentRasterDeleter(PersistentProcessing):
-
-    def __init__(self, *args):
-        PersistentProcessing.__init__(self, *args)
-
-    def _execute(self):
-        """Delete a specific raster layer
-
-        Use the original mapset for processing
-        """
-        self._setup()
-
-        raster_name = self.map_name
-        self.required_mapsets.append(self.target_mapset_name)
-
-        pc = {}
-        pc["1"] = {"module": "g.remove", "inputs": {"type": "raster",
-                                                    "name": raster_name},
-                   "flags": "f"}
-
-        self.skip_region_check = True
-        process_list = self._validate_process_chain(process_chain=pc,
-                                                    skip_permission_check=True)
-        self._check_lock_target_mapset()
-        self._create_temp_database(self.required_mapsets)
-        self._create_grass_environment(grass_data_base=self.temp_grass_data_base,
-                                       mapset_name=self.target_mapset_name)
-
-        self._execute_process_list(process_list)
-
-        if "WARNING: No data base element files found" in "\n".join(
-                self.module_output_log[0]["stderr"]):
-            raise AsyncProcessError("Raster layer <%s> not found" % raster_name)
-
-        self.finish_message = "Raster layer <%s> successfully removed." % raster_name
-
-
-def start_create_job(*args):
-    processing = PersistentRasterCreator(*args)
-    processing.run()
-
-
-class PersistentRasterCreator(PersistentProcessing):
-
-    def __init__(self, *args):
-
-        PersistentProcessing.__init__(self, *args)
-
-    def _execute(self):
-        """Create a specific raster layer
-
-        This approach is complex, since the raster generation is performed in a local
-        temporary mapset that is later merged into the target mapset. Workflow:
-
-        1. Check the process chain
-        2. Lock the temp and target mapsets
-        3. Setup GRASS and create the temporary mapset
-        4. Execute g.list of the first process chain to check if the target
-           raster exists
-        5. If the target raster does not exists then run r.import
-        6. Copy the local temporary mapset to the storage and merge it into the
-           target mapset
-        """
-        self._setup()
-
-        raster_name = self.map_name
-        self.required_mapsets.append(self.target_mapset_name)
-
-        pc_1 = {}
-        pc_1["1"] = {"module": "g.list", "inputs": {
-            "type": "raster",
-            "pattern": raster_name,
-            "mapset": self.target_mapset_name}}
-        # Check the first process chain
-        pc_1 = self._validate_process_chain(skip_permission_check=True,
-                                            process_chain=pc_1)
-
-        pc_2 = {}
-        pc_2["1"] = {
-            "module": "r.import",
-            "inputs": {
-                "input": self.rdc.request_data,
-                "output": raster_name
-                }
-        }
-        # Check the second process chain
-        pc_2 = self._validate_process_chain(skip_permission_check=True,
-                                            process_chain=pc_2)
-
-        self._check_lock_target_mapset()
-        self._lock_temp_mapset()
-        self._create_temporary_grass_environment()
-        self._execute_process_list(pc_1)
-
-        # check if raster exists
-        raster_list = self.module_output_log[0]["stdout"].split("\n")
-
-        if len(raster_list[0]) > 0:
-            try:
-                os.remove(self.rdc.request_data)
-            except Exception:
-                pass
-            raise AsyncProcessError(
-                "Raster layer <%s> exists. Please rename it or delete the old "
-                "raster layer" % raster_name)
-
-        self._execute_process_list(pc_2)
-        self._copy_merge_tmp_mapset_to_target_mapset()
-
-        # Delete imported file
-        msg = ""
-        try:
-            os.remove(self.rdc.request_data)
-        except Exception:
-            msg = " WARNING: Uploaded file can not be removed."
-
-        self.finish_message = (f"Raster layer <{raster_name}> successfully "
-                               f"imported.{msg}")
